+++ app/src/androidTest/java/woowacourse/movie/ReservationCompleteActivityTest.kt
@@ -12,7 +12,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import woowacourse.movie.activity.ReservationCompleteActivity
+import woowacourse.movie.activity.reservation.ReservationCompleteActivity
 import woowacourse.movie.domain.BookingStatus
 import woowacourse.movie.domain.MemberCount
 import woowacourse.movie.dto.ReservationDto
