--- app/src/main/java/woowacourse/movie/activity/ReservationSeatActivity.kt
@@ -1,26 +1,0 @@
-package woowacourse.movie.activity
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import woowacourse.movie.R
-import woowacourse.movie.databinding.ActivitySeatBinding
-
-class ReservationSeatActivity : AppCompatActivity() {
-    private val binding: ActivitySeatBinding by lazy {
-        ActivitySeatBinding.inflate(layoutInflater)
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(binding.root)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.booking)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
-    }
-}
