+++ app/src/main/java/woowacourse/movie/activity/reservation/ReservationSeatActivity.kt
@@ -1,0 +1,63 @@
+package woowacourse.movie.activity.reservation
+
+import android.app.AlertDialog
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import woowacourse.movie.R
+import woowacourse.movie.databinding.ActivitySeatBinding
+import woowacourse.movie.dto.ReservationDto
+
+class ReservationSeatActivity : AppCompatActivity() {
+    private val binding: ActivitySeatBinding by lazy {
+        ActivitySeatBinding.inflate(layoutInflater)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContentView(binding.root)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.seat)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+    }
+
+    companion object {
+        private const val BOOKING_STATUS_KEY = "bookingStatus"
+
+        fun newIntent(
+            from: Context,
+            dto: ReservationDto,
+        ): Intent {
+            return Intent(from, ReservationSeatActivity::class.java).apply {
+                putExtra(BOOKING_STATUS_KEY, dto)
+            }
+        }
+    }
+}
+
+private class DialogBuilder(val context: Context) {
+    private var dialog: AlertDialog.Builder =
+        AlertDialog.Builder(context)
+            .setTitle(context.getString(R.string.complete_dialog_title))
+            .setMessage(context.getString(R.string.complete_dialog_message))
+            .setNegativeButton(R.string.complete_dialog_negative_button) { dialog, _ ->
+                dialog.dismiss()
+            }
+            .setCancelable(false)
+
+    fun onPositiveButtonClicked(block: () -> Unit): DialogBuilder {
+        dialog.setPositiveButton(context.getString(R.string.complete_dialog_positive_button)) { _, _ ->
+            block()
+        }
+        return this
+    }
+
+    fun build(): AlertDialog = dialog.create()
+}
