+++ app/src/main/res/layout/activity_reservation_complete.xml
@@ -5,7 +5,7 @@
     android:id="@+id/booking_success"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.MainActivity">
+    tools:context=".activity.main.MainActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notice_text_container"
